requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.2
Django==5.1.7
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-ninja==1.3.0
django-ninja-extra==0.22.6
django-ninja-jwt==5.3.5
django-timescaledb==0.2.13
dnspython==2.7.0
email_validator==2.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
injector==0.22.0
packaging==24.2
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.0
whitenoise==6.9.0
